+++ lib/lotus/loader.rb
@@ -40,6 +40,7 @@ module Lotus
       _configure_model_framework! if defined?(Lotus::Model)
       _configure_controller_framework!
       _configure_view_framework!
+      _configure_logger!
     end
 
     def _configure_controller_framework!
@@ -87,6 +88,12 @@ module Lotus
       end
     end
 
+    def _configure_logger!
+      unless application_module.const_defined?('Logger', false)
+        logger = Lotus::Logger.new(application_module.to_s)
+        application_module.const_set('Logger', logger)
+      end
+    end
 
     def load_frameworks!
       _load_view_framework!